advanced-react ch37: lock down delete button

+++ advanced-react/sick-fits/backend/src/resolvers/Mutation.js
@@ -34,8 +34,17 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
-    const item = await ctx.db.query.item({ where }, '{ id title }')
-    // TODO: check if user own the item, or has permission to delete it
+    const item = await ctx.db.query.item({ where }, '{ id, title, user { id } }')
+
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(
+      permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+    )
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You're not allowed.")
+    }
+
     return ctx.db.mutation.deleteItem({ where }, info)
   },
   async signup(parent, args, ctx, info) {
